% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_funs.R
\name{binary_collapse}
\alias{binary_collapse}
\title{Collapse Grouped Binary Data}
\usage{
binary_collapse(data, outcome)
}
\arguments{
\item{data}{the data frame containing binary data}

\item{outcome}{the binary response variable}
}
\value{
a data frame in the binomial format
}
\description{
This function can be used to take binary data and collapse it
into binomial data. Occassionally useful.
}
