% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_funs.R
\name{binary_gof}
\alias{binary_gof}
\title{Binary Goodness of Fit Test}
\usage{
binary_gof(fit, ngrp = 10, print.table = TRUE)
}
\arguments{
\item{fit}{a model object fir with \code{glm()}}

\item{ngrp}{the number of groups for the binary test. Defaults to 10.}

\item{print.table}{Logical, defaults to \code{TRUE}. Prints results as well as
returning them.}
}
\value{
test results
}
\description{
Binary Goodness of Fit Test
}
