% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary.R
\name{categorical_summary}
\alias{categorical_summary}
\title{Categorical Summary}
\usage{
categorical_summary(values, level)
}
\arguments{
\item{values}{a vector (usually from a data frame)}

\item{level}{the level to summarize}
}
\value{
a character string listing the number of occurrences of \code{level} in
\code{values} along with the percentage in parentheses
}
\description{
Summarize a categorical variable (just a specific level) and give the
number of observations and the corresponding percentage.
}
