% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_summary.R
\name{continuous_summary}
\alias{continuous_summary}
\title{Continuous Value Summary}
\usage{
continuous_summary(
  values,
  subset = NULL,
  type = "mean",
  accuracy = 0.01,
  na.rm = TRUE
)
}
\arguments{
\item{values}{the values to be summarized, a numeric vector}

\item{subset}{a logical vector denoting the subset of \code{values}
that you wish to summarize. Defaults to \code{NULL}.}

\item{type}{either \code{"mean"} in which case the mean and standard
deviation are computed, or \code{"median"} where the median and IQR
are computed.}

\item{accuracy}{the rounding accuracy of the output string, as in
\code{scales::number()}.}

\item{na.rm}{whether or not to remove \code{NA}s, defaults to \code{TRUE}}
}
\value{
a character string summarizing \code{values}
}
\description{
Continuous Value Summary
}
