% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ci.R
\name{glm_ci}
\alias{glm_ci}
\title{Generalized Linear Model Confidence Interval}
\usage{
glm_ci(model, transform = TRUE, robust = FALSE)
}
\arguments{
\item{model}{a model object created by \code{glm()}}

\item{transform}{a logical indicating whether or not to transform the
coefficients and other estimates via the link function. Defaults to \code{TRUE}.}

\item{robust}{a logical indicating whether or not to use the robust
(sandwich) variance estimator when computing the standard errors
of the parameter estimates. Defaults to \code{FALSE}.}
}
\value{
a matrix containing the parameter estimates, confidence
interval, test statistic, and corresponding t-value.
}
\description{
Due to the shameful lack of a confidence interval in base \code{R} \code{glm()} output,
this function was created to take a model and produce a summary table of
coefficients including confidence interval estimates.
}
