% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_utils.R
\name{ifelse1}
\alias{ifelse1}
\title{Alternative ifelse Function}
\usage{
ifelse1(test, yes, no)
}
\arguments{
\item{test}{logical test}

\item{yes}{return if \code{test} is true}

\item{no}{return if \code{test} is false}
}
\value{
either \code{yes} or \code{no} depending on the value of \code{test}
}
\description{
Alternative ifelse Function
}
