% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbl_helpers.R
\name{kbl_pdf}
\alias{kbl_pdf}
\title{Kable PDF Format}
\usage{
kbl_pdf(df, colnames, caption, escape = TRUE)
}
\arguments{
\item{df}{data frame to be output in the \code{kbl}}

\item{colnames}{the displayed column names in the output}

\item{caption}{the caption of the table in the output}

\item{escape}{logical, whether or not to escape the table.
Defaults to \code{TRUE} (i.e. math notation will not be formatted).}
}
\value{
an object with class \code{kbl}
}
\description{
Presets for table formats I commonly use in my reports.
}
