% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ci.R
\name{lm_ci}
\alias{lm_ci}
\title{Linear Model Confidence Interval}
\usage{
lm_ci(model, expcoef = FALSE, robust = FALSE)
}
\arguments{
\item{model}{a model object created by \code{lm()}}

\item{expcoef}{a logical indicating whether or not to exponentiate the
coefficients and other estimates. Generally used if the model has a
log-transformed response. Defaults to \code{FALSE}.}

\item{robust}{a logical indicating whether or not to use the robust
(sandwich) variance estimator when computing the standard errors
of the parameter estimates. Defaults to \code{FALSE}.}
}
\value{
a matrix containing the parameter estimates, confidence
interval, test statistic, and corresponding t-value.
}
\description{
Due to the shameful lack of a confidence interval in base \code{R} \code{lm()} output,
this function was created to take a model and produce a summary table of
coefficients including confidence interval estimates.
}
