% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_funs.R
\name{multinom_summ}
\alias{multinom_summ}
\title{Summarize a multinomial Model Fit}
\usage{
multinom_summ(model)
}
\arguments{
\item{model}{a multinomial model fit with \code{nnet::multinom()}}
}
\value{
a summary of the multinomial model fit
}
\description{
This function summarizes each level of a multinomial model
fit as compared to the first level.
}
