% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_covariance.R
\name{robust.se.glm}
\alias{robust.se.glm}
\title{Robust standard error computation for GLM}
\usage{
robust.se.glm(glm.obj)
}
\arguments{
\item{model}{a model object created by \code{glm()}}
}
\value{
a matrix containing the robust standard errors and confidence intervals
}
\description{
Robust standard error computation for GLM
}
