% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_covariance.R
\name{robust.se.lm}
\alias{robust.se.lm}
\title{Robust standard error computation for LM}
\usage{
robust.se.lm(model)
}
\arguments{
\item{model}{a model object created by \code{lm()}}
}
\value{
a matrix containing the robust standard errors and confidence intervals
}
\description{
Robust standard error computation for LM
}
