% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_covariance.R
\name{robust.vcov.glm}
\alias{robust.vcov.glm}
\title{Robust Variance computation for GLMs}
\usage{
robust.vcov.glm(glm.obj)
}
\arguments{
\item{glm.obj}{a model created by \code{glm()}}
}
\value{
robust covariance matrix
}
\description{
Robust Variance computation for GLMs
}
