% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_covariance.R
\name{robust.vcov.lm}
\alias{robust.vcov.lm}
\title{Robust variance computation for LM}
\usage{
robust.vcov.lm(lm.obj)
}
\arguments{
\item{lm.obj}{a model created by \code{lm()}}
}
\value{
the robust covariance matrix
}
\description{
Robust variance computation for LM
}
