% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_helpers.R
\name{theme_spencer}
\alias{theme_spencer}
\title{Spencer's Preferred Theme}
\usage{
theme_spencer()
}
\value{
a ggplot \code{theme()} object
}
\description{
Subject to change at any given point in time without warning.
}
